# fix(perf): improve ValueDB lookup performance