--- a/test/valuedb-perf.ts
+++ b/test/valuedb-perf.ts
@@ -15,6 +15,7 @@ const valueDBs = new Map<number, ValueDB>();
 	await values.open();
 	await metadata.open();
 	// add a shitton of values
+	console.time("create value DB");
 	const MAX_NODES = 100;
 	for (let nodeId = 1; nodeId <= MAX_NODES; nodeId++) {
 		const valueDB = new ValueDB(nodeId, values, metadata);
@@ -34,6 +35,19 @@ const valueDBs = new Map<number, ValueDB>();
 			}
 		}
 	}
+	await values.close();
+	console.timeEnd("create value DB");
+
+	console.time("open value DB");
+	await values.open();
+	console.timeEnd("open value DB");
+
+	console.time("index value DB");
+	for (let nodeId = 1; nodeId <= MAX_NODES; nodeId++) {
+		const valueDB = new ValueDB(nodeId, values, metadata);
+		valueDBs.set(nodeId, valueDB);
+	}
+	console.timeEnd("index value DB");
 
 	console.log(`db contains ${values.size} values`);
 	const numRounds = 100;
@@ -41,11 +55,13 @@ const valueDBs = new Map<number, ValueDB>();
 	console.log(`calling getValues ${numRounds} times`);
 
 	let start = highResTimestamp();
+	let numValues: number;
 	for (let i = 1; i <= numRounds; i++) {
-		valueDBs.get(1)!.getValues(5);
+		numValues = valueDBs.get(1)!.getValues(5).length;
 	}
 	let duration = highResTimestamp() - start;
 	console.log(`took ${(duration / 1e6 / numRounds).toFixed(2)} ms / call`);
+	console.log(`found ${numValues!} values`);
 
 	console.log(`calling clear for each node (${MAX_NODES})`);