This PR attempts to fix the performance issues found in #1493. The Value DB is deliberately huge (550000 values), so we can judge the impact better.

## Step 1:
`dbKeyToValueIdFast` is optimized, but still too slow to be in the hot path. We now filter values using string comparison instead of parsing JSON first and likely discarding the result afterwards.

**Before:**
```
db contains 550000 values
calling getValues 100 times
took 550.62 ms / call
```

**After:**
```
db contains 550000 values
calling getValues 100 times
took 44.97 ms / call
```

**total speedup:** 12x

## Step 2:
The `clear()` methods were parsing the DB keys although the result is only used conditionally. That no longer happens now.
**Before:**
```
calling clear for each node (100)
took 106.55 ms / call
```

**After:**
```
calling clear for each node (100)
took 97.88 ms / call
```

## Step 3:
`indexOf` turns out to be faster, but only if we include the leading `{` and `,`.
**After:**
```
calling getValues 100 times
took 39.94 ms / call
calling clear for each node (100)
took 95.81 ms / call
```

**total speedup:** 13.8x

## Step 4:
Keeping track of the own IDs completely avoids looping altogether. However this adds an upfront cost for the index creation:
**Before:**
```
index value DB: 0.162ms (NO-OP)

calling getValues 100 times
took 40.04 ms / call

calling clear for each node (100)
took 91.89 ms / call
```

**After:**
```
index value DB: 6770.065ms

calling getValues 100 times
took 1.23 ms / call

calling clear for each node (100)
took 12.35 ms / call
```

**total speedup:** 🔥 447x (ignoring the upfront cost)

## Step 5: TODO: Speed up index generation
**After:**
```
index value DB: 122.762ms

calling getValues 100 times
took 1.23 ms / call

calling clear for each node (100)
took 12.31 ms / call
```

**total speedup:** 🔥 447x (with negligible upfront cost 🎉)

---

/cc @marcelveldt
fixes: #1493