@robertsLando 2021-01-26
@AlCalzone indexOf is even faster: https://jsben.ch/om6Xb (around 20%)

Anyway I think that this could be even faster with a cache