@AlCalzone 2021-01-26
> indexOf is even faster

Gave it a try 👎🏻:
```
calling getValues 10 times
took 1362.05 ms / call
```