@robertsLando 2021-01-26
> Gave it a try 👎🏻:

Lol maybe that's only for browsers :man_shrugging: