@AlCalzone 2021-01-26
Nevermind, I messed up.
```
db contains 550000 values
calling getValues 100 times
took 41.28 ms / call
```