@robertsLando 2021-01-26
@AlCalzone Personally I think that the build index time is really too high, I think that the step 3 is enought