@AlCalzone 2021-01-26
Index generation for the 550k values DB is down to ~123 ms, so all concerns should be addressed.

@marcelveldt I'll ask the bot to create a test build so you can check this out.
@zwave-js-bot pack this.