@marcelveldt 2021-01-26
I gave this version a testdrive and it's REALLY amazing. That nasty dumpState call in the server took 7+ seconds on my main mesh yesterday (without this fix) and this now takes about 50ms :-)

Also can confirm that it didn't break anything else I could discover.